+++ src/main/java/org/organicdesign/fp/function/Function2.java
@@ -13,12 +13,11 @@
 // limitations under the License.
 package org.organicdesign.fp.function;
 
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.BiFunction;
 
-import org.organicdesign.fp.Option;
+import org.organicdesign.fp.oneOf.Option;
 import org.organicdesign.fp.tuple.Tuple2;
 
 /**